% Auto-generated: do not edit by hand
\name{dashWindowCloseListener}

\alias{dashWindowCloseListener}

\title{DashWindowCloseListener component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
dashWindowCloseListener(id=NULL, status=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{status}{Character. The status of the window. Mounted or closing.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

